requirements.txt
scikit-learn
matplotlib
nltk
numpy
PyPDF2
pdfminer.six
requests
